Update issue templates
My initial creation of three issue templates for the repository. The most used one will be the default issue template for starters. Perhaps will try the Feature Request and Bug Report. A possible future template would be for Epic and Dependent issues. Or these could be listed as optional sections in default issue template.